tools/picturas-text-extraction-tool-ms/pyproject.toml
[tool.poetry]
name = "picturas-text-extractor-tool-ms"
version = "0.1.0"
description = "PictuRAS microservice (tool) that applies a text extractor to an input image, generating a text output present on image."
authors = ["Pedro Viana <pg57895@alunos.uminho.pt>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
pika = "^1.3.2"
pydantic = "^2.10.3"
pillow = "^11.0.0"
pytesseract = "^0.3.13"



[tool.poetry.group.dev.dependencies]
yapf = "^0.43.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
